logo mockup
A quick idea I had for a logo. Basically trying to convey the idea that we're increasing visibility of people previously "hidden" within the scientific community (via internet networking). Not really sure if the idea comes across at all. If we like this idea, I can spend more time fleshing out the actual design and details.