Created homepage using Yeoman.
I installed and created the main scaffolding for the app using Yeoman. When deciding on configuration, I used Angular for our front-end framework. For CSS framework, I used Bootstrap. I did not use SASS or LESS, though that may be added later. Using Angular, I created the views for each link on the homepage. I also created a few controllers as a starting point for when we want to use more dynamic data and utilize Angular's data-binding capabilities later.